consolidate base and core
